backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
pgvector==0.3.5
asyncpg==0.30.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI Providers
openai==1.51.0
anthropic==0.36.2
google-generativeai==0.8.3
ollama==0.3.3

# Document Processing
pypdf2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
tiktoken==0.8.0

# Embeddings & Vector
numpy==1.26.4
sentence-transformers==3.2.1

# Cache & Performance
redis==5.1.1
aiofiles==24.1.0

# Utils
httpx==0.27.2
tenacity==9.0.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
structlog==24.4.0
